dag 环检查算法

+++ src/main/java/com/hancher/sentinel/enums/DagNodeEnum.java
@@ -22,5 +22,5 @@ public enum DagNodeEnum {
     end(-1),
     ;
 
-    private final int code;
+    private final long code;
 }
